\name{read_MESSAGE}
\alias{read_MESSAGE}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Read MESSAGE output variables as dataframes
}
\description{
Read MESSAGE output GDX files, using the rgdx.var function and gdxrrw
}
\usage{
read_MESSAGE(msg_version, msg_variable)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{msg_version}{
Version of MESSAGE scenario (e.g. 70)
}
  \item{msg_variable}{
Output variable that you would like to read (e.g. 'ACT')
}
}

\author{
Jun Shepard
}

\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
